started new architecture
